translation-server/requirements.txt
python-dotenv==1.0.1
fastapi[standard]==0.115.0
requests==2.32.3
deepgram-sdk==3.7.0